LoudnessMetadataScanner now more cleanly warns upon IOExceptions encountered when trying to load metadata files.
The old code would have reported this as an error with a logged stack trace, but since we know this to be a safe situation we now just log a warning.

This can occur on rare occasion when the foreground thread is trying to load the metadata file at the exact same time as BS1770GAIN is trying to save it.