Fixed a number of (small) memory leaks, most related to dynamic text and dummy 1x1 bitmaps.
Note that I've started shifting toward usage of the standard Dispose method, as per .net best practice. We should continue to push the code in this direction as we work, because the old-style release methods hide the fact that Dispose is used underneath, defeat tools which can help find code that fails to Dispose of objects, and does not integrate properly into exception handling the way the the using statement does.